backend/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
futures==3.1.1
goslate==1.5.2
idna==3.3
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyDictionary==2.0.1
python-decouple==3.4
pytz==2021.1
requests==2.27.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tzdata==2021.5
urllib3==1.26.5
whitenoise==5.3.0
gunicorn==20.0.4
PyDictionary==2.0.1